requirements.txt
python-jobspy
python-telegram-bot==20.7
apscheduler
anthropic
python-dotenv
